+++ src/Smartstore.Web/Areas/Admin/Models/Settings/MediaSettingsModel.cs
@@ -123,6 +123,9 @@ namespace Smartstore.Admin.Models
 
         public QuantizationMethod GifQuantizationMethod { get; set; }
 
+        [LocalizedDisplay("*OffloadEmbeddedImagesOnSave")]
+        public bool OffloadEmbeddedImagesOnSave { get; set; }
+
         #endregion
 
         #region Response Caching